- allow selecting keyring directory and user - make it possible to generate multiple keys in the same keyring
git-svn-id: svn+ssh://svn.mageia.org/svn/adm/puppet@941 db665093-e049-48a9-b3c2-13f1584ce46d

+++ modules/gnupg/templates/create_gnupg_keys.sh
@@ -1,11 +1,24 @@
 #!/bin/bash
 
-NAME=$1
+BATCHFILE="$1"
+HOMEDIR="$2"
+LOCK="$3"
+
+test $# -eq 3 || exit 1
+
+if [ -e "$LOCK" ]
+then
+    echo "Lock file already exist." 1>&2
+    echo "Remove $LOCK if you want to regenerate key." 1>&2
+    exit 2
+fi
+
+touch "$LOCK"
 
 /sbin/rngd -f -r /dev/urandom &
 RAND=$!
-cd /etc/gnupg/keys/
-gpg --homedir /etc/gnupg/keys/ --batch --gen-key /etc/gnupg/batches/$NAME.batch 
+cd $HOMEDIR
+gpg --homedir $HOMEDIR --batch --gen-key $BATCHFILE
 EXIT=$?
 
 kill $RAND
